research/requirements.txt
cassandra-driver==3.26.0
pymongo==4.3.3
pydantic==1.10.7
